refactor eval runtime modules
