+++ crates/klassic-eval/src/lib.rs
@@ -11,153 +11,21 @@ use std::time::{Duration, Instant};
 use klassic_rewrite::rewrite_expression;
 use klassic_span::{Diagnostic, DiagnosticKind, Severity, SourceFile, Span};
 use klassic_syntax::{
-    BinaryOp, Expr, RecordField, TypeAnnotation, TypeClassConstraint, UnaryOp,
-    parse_inline_expression, parse_source,
+    BinaryOp, Expr, RecordField, TypeAnnotation, TypeClassConstraint, parse_inline_expression,
+    parse_source,
 };
 use klassic_types::{
     KnownType, clear_user_binding_types, clear_user_typeclass_infos,
     typecheck_program_with_bindings,
 };
 
-#[derive(Clone, Debug)]
-pub enum Value {
-    Int(i64),
-    Long(i64),
-    Float(f32),
-    Double(f64),
-    Bool(bool),
-    String(String),
-    Null,
-    Unit,
-    List(Vec<Value>),
-    Map(Vec<(Value, Value)>),
-    Set(Vec<Value>),
-    Record {
-        name: String,
-        fields: Vec<(String, Value)>,
-    },
-    TypeClassMethod(String),
-    BoundTypeClassMethod {
-        name: String,
-        fallback: Box<Value>,
-    },
-    BuiltinFunction(Rc<BuiltinFunctionValue>),
-    Function(Rc<FunctionValue>),
-}
+mod builtin_registry;
+mod ops;
+mod value;
 
-impl PartialEq for Value {
-    fn eq(&self, other: &Self) -> bool {
-        match (self, other) {
-            (Self::Int(lhs), Self::Int(rhs)) => lhs == rhs,
-            (Self::Long(lhs), Self::Long(rhs)) => lhs == rhs,
-            (Self::Float(lhs), Self::Float(rhs)) => lhs == rhs,
-            (Self::Double(lhs), Self::Double(rhs)) => lhs == rhs,
-            (Self::Bool(lhs), Self::Bool(rhs)) => lhs == rhs,
-            (Self::String(lhs), Self::String(rhs)) => lhs == rhs,
-            (Self::Null, Self::Null) => true,
-            (Self::Unit, Self::Unit) => true,
-            (Self::List(lhs), Self::List(rhs)) => lhs == rhs,
-            (Self::Map(lhs), Self::Map(rhs)) => lhs == rhs,
-            (Self::Set(lhs), Self::Set(rhs)) => lhs == rhs,
-            (
-                Self::Record {
-                    name: lhs_name,
-                    fields: lhs_fields,
-                },
-                Self::Record {
-                    name: rhs_name,
-                    fields: rhs_fields,
-                },
-            ) => lhs_name == rhs_name && lhs_fields == rhs_fields,
-            (Self::Int(lhs), Self::Double(rhs)) => (*lhs as f64) == *rhs,
-            (Self::Double(lhs), Self::Int(rhs)) => *lhs == (*rhs as f64),
-            (Self::Int(lhs), Self::Long(rhs)) => *lhs == *rhs,
-            (Self::Long(lhs), Self::Int(rhs)) => *lhs == *rhs,
-            (Self::Int(lhs), Self::Float(rhs)) => (*lhs as f32) == *rhs,
-            (Self::Float(lhs), Self::Int(rhs)) => *lhs == (*rhs as f32),
-            (Self::Long(lhs), Self::Double(rhs)) => (*lhs as f64) == *rhs,
-            (Self::Double(lhs), Self::Long(rhs)) => *lhs == (*rhs as f64),
-            (Self::Long(lhs), Self::Float(rhs)) => (*lhs as f32) == *rhs,
-            (Self::Float(lhs), Self::Long(rhs)) => *lhs == (*rhs as f32),
-            (Self::Float(lhs), Self::Double(rhs)) => (*lhs as f64) == *rhs,
-            (Self::Double(lhs), Self::Float(rhs)) => *lhs == (*rhs as f64),
-            _ => false,
-        }
-    }
-}
-
-impl fmt::Display for Value {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Int(value) => write!(f, "{value}"),
-            Self::Long(value) => write!(f, "{value}"),
-            Self::Float(value) => {
-                if value.fract() == 0.0 {
-                    write!(f, "{value:.1}")
-                } else {
-                    write!(f, "{value}")
-                }
-            }
-            Self::Double(value) => {
-                if value.fract() == 0.0 {
-                    write!(f, "{value:.1}")
-                } else {
-                    write!(f, "{value}")
-                }
-            }
-            Self::Bool(value) => write!(f, "{value}"),
-            Self::String(value) => write!(f, "{value}"),
-            Self::Null => write!(f, "null"),
-            Self::Unit => write!(f, "()"),
-            Self::List(values) => {
-                write!(f, "[")?;
-                for (index, value) in values.iter().enumerate() {
-                    if index > 0 {
-                        write!(f, ", ")?;
-                    }
-                    write!(f, "{value}")?;
-                }
-                write!(f, "]")
-            }
-            Self::Map(entries) => {
-                write!(f, "%[")?;
-                for (index, (key, value)) in entries.iter().enumerate() {
-                    if index > 0 {
-                        write!(f, ", ")?;
-                    }
-                    write!(f, "{key}: {value}")?;
-                }
-                write!(f, "]")
-            }
-            Self::Set(values) => {
-                write!(f, "%(")?;
-                for (index, value) in values.iter().enumerate() {
-                    if index > 0 {
-                        write!(f, ", ")?;
-                    }
-                    write!(f, "{value}")?;
-                }
-                write!(f, ")")
-            }
-            Self::Record { name, fields } => {
-                write!(f, "#{name}(")?;
-                for (index, (_, value)) in fields.iter().enumerate() {
-                    if index > 0 {
-                        write!(f, ", ")?;
-                    }
-                    write!(f, "{value}")?;
-                }
-                write!(f, ")")
-            }
-            Self::TypeClassMethod(name) => write!(f, "<typeclass-method:{name}>"),
-            Self::BoundTypeClassMethod { name, .. } => {
-                write!(f, "<bound-typeclass-method:{name}>")
-            }
-            Self::BuiltinFunction(function) => write!(f, "<builtin:{}>", function.name),
-            Self::Function(_) => write!(f, "<function>"),
-        }
-    }
-}
+use builtin_registry::{builtin_arity, builtin_name, value_method_builtin_name};
+use ops::{eval_binary, eval_unary};
+pub use value::{BuiltinFunctionValue, FunctionValue, Value};
 
 #[derive(Clone, Debug)]
 enum ThreadValueSnapshot {
@@ -229,21 +97,6 @@ struct ThreadInstanceDictionaryEntrySnapshot {
 type ThreadModuleExports = HashMap<String, ThreadValueSnapshot>;
 type ThreadInstanceMethods = HashMap<String, Vec<ThreadInstanceMethodEntrySnapshot>>;
 type ThreadInstanceDictionaries = HashMap<String, Vec<ThreadInstanceDictionaryEntrySnapshot>>;
-
-#[derive(Clone, Debug)]
-pub struct BuiltinFunctionValue {
-    name: &'static str,
-    bound_args: Vec<Value>,
-}
-
-#[derive(Clone, Debug)]
-pub struct FunctionValue {
-    params: Vec<String>,
-    param_annotations: Vec<Option<TypeAnnotation>>,
-    constraints: Vec<TypeClassConstraint>,
-    body: Expr,
-    env: Environment,
-}
 
 #[derive(Clone, Debug)]
 pub struct EvaluationError {
@@ -1962,131 +1815,6 @@ fn clear_user_instances() {
     USER_INSTANCE_DICTIONARIES.with(|dictionaries| dictionaries.borrow_mut().clear());
 }
 
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-enum NumericKind {
-    Int,
-    Long,
-    Float,
-    Double,
-}
-
-#[derive(Clone, Copy, Debug)]
-enum NumericValue {
-    Int(i64),
-    Long(i64),
-    Float(f32),
-    Double(f64),
-}
-
-impl NumericValue {
-    fn kind(self) -> NumericKind {
-        match self {
-            Self::Int(_) => NumericKind::Int,
-            Self::Long(_) => NumericKind::Long,
-            Self::Float(_) => NumericKind::Float,
-            Self::Double(_) => NumericKind::Double,
-        }
-    }
-
-    fn promote(self, kind: NumericKind) -> Self {
-        match kind {
-            NumericKind::Int => match self {
-                Self::Int(value) => Self::Int(value),
-                _ => unreachable!("promotion to Int should only happen for Int"),
-            },
-            NumericKind::Long => match self {
-                Self::Int(value) => Self::Long(value),
-                Self::Long(value) => Self::Long(value),
-                _ => unreachable!("promotion to Long should only happen for integral types"),
-            },
-            NumericKind::Float => match self {
-                Self::Int(value) => Self::Float(value as f32),
-                Self::Long(value) => Self::Float(value as f32),
-                Self::Float(value) => Self::Float(value),
-                _ => unreachable!("promotion to Float should only happen for numeric types"),
-            },
-            NumericKind::Double => match self {
-                Self::Int(value) => Self::Double(value as f64),
-                Self::Long(value) => Self::Double(value as f64),
-                Self::Float(value) => Self::Double(value as f64),
-                Self::Double(value) => Self::Double(value),
-            },
-        }
-    }
-
-    fn as_f64(self) -> f64 {
-        match self {
-            Self::Int(value) => value as f64,
-            Self::Long(value) => value as f64,
-            Self::Float(value) => value as f64,
-            Self::Double(value) => value,
-        }
-    }
-
-    fn into_value(self) -> Value {
-        match self {
-            Self::Int(value) => Value::Int(value),
-            Self::Long(value) => Value::Long(value),
-            Self::Float(value) => Value::Float(value),
-            Self::Double(value) => Value::Double(value),
-        }
-    }
-}
-
-fn numeric_value(value: &Value) -> Option<NumericValue> {
-    match value {
-        Value::Int(value) => Some(NumericValue::Int(*value)),
-        Value::Long(value) => Some(NumericValue::Long(*value)),
-        Value::Float(value) => Some(NumericValue::Float(*value)),
-        Value::Double(value) => Some(NumericValue::Double(*value)),
-        _ => None,
-    }
-}
-
-fn promote_numeric_pair(
-    lhs: NumericValue,
-    rhs: NumericValue,
-) -> (NumericValue, NumericValue, NumericKind) {
-    use NumericKind::*;
-    let rank = |kind| match kind {
-        Int => 0,
-        Long => 1,
-        Float => 2,
-        Double => 3,
-    };
-    let kind = if rank(lhs.kind()) >= rank(rhs.kind()) {
-        lhs.kind()
-    } else {
-        rhs.kind()
-    };
-    (lhs.promote(kind), rhs.promote(kind), kind)
-}
-
-fn eval_unary(op: UnaryOp, value: Value, span: Span) -> Result<Value, Diagnostic> {
-    match (op, value) {
-        (UnaryOp::Plus, Value::Int(value)) => Ok(Value::Int(value)),
-        (UnaryOp::Plus, Value::Long(value)) => Ok(Value::Long(value)),
-        (UnaryOp::Plus, Value::Float(value)) => Ok(Value::Float(value)),
-        (UnaryOp::Plus, Value::Double(value)) => Ok(Value::Double(value)),
-        (UnaryOp::Minus, Value::Int(value)) => value
-            .checked_neg()
-            .map(Value::Int)
-            .ok_or_else(|| Diagnostic::runtime(span, "integer overflow in unary negation")),
-        (UnaryOp::Minus, Value::Long(value)) => value
-            .checked_neg()
-            .map(Value::Long)
-            .ok_or_else(|| Diagnostic::runtime(span, "integer overflow in unary negation")),
-        (UnaryOp::Minus, Value::Float(value)) => Ok(Value::Float(-value)),
-        (UnaryOp::Minus, Value::Double(value)) => Ok(Value::Double(-value)),
-        (UnaryOp::Not, Value::Bool(value)) => Ok(Value::Bool(!value)),
-        (UnaryOp::Plus, _) | (UnaryOp::Minus, _) => Err(Diagnostic::runtime(
-            span,
-            "numeric unary operator expected a number",
-        )),
-        (UnaryOp::Not, _) => Err(Diagnostic::runtime(span, "`!` expected a boolean")),
-    }
-}
-
 fn eval_logical(
     op: BinaryOp,
     lhs: &Expr,
@@ -2118,196 +1846,6 @@ fn eval_logical(
             "logical operators require booleans",
         )),
     }
-}
-
-fn eval_binary(op: BinaryOp, lhs: Value, rhs: Value, span: Span) -> Result<Value, Diagnostic> {
-    match op {
-        BinaryOp::Add => eval_add(lhs, rhs, span),
-        BinaryOp::Subtract => eval_subtract(lhs, rhs, span),
-        BinaryOp::Multiply => eval_multiply(lhs, rhs, span),
-        BinaryOp::Divide => eval_divide(lhs, rhs, span),
-        BinaryOp::Less | BinaryOp::LessEqual | BinaryOp::Greater | BinaryOp::GreaterEqual => {
-            eval_comparison(op, lhs, rhs, span)
-        }
-        BinaryOp::Equal | BinaryOp::NotEqual => {
-            let equal = lhs == rhs;
-            if matches!(op, BinaryOp::Equal) {
-                Ok(Value::Bool(equal))
-            } else {
-                Ok(Value::Bool(!equal))
-            }
-        }
-        BinaryOp::BitAnd | BinaryOp::BitOr | BinaryOp::BitXor => eval_bitwise(op, lhs, rhs, span),
-        BinaryOp::LogicalAnd | BinaryOp::LogicalOr => unreachable!("handled separately"),
-    }
-}
-
-fn eval_add(lhs: Value, rhs: Value, span: Span) -> Result<Value, Diagnostic> {
-    match (lhs, rhs) {
-        (Value::String(lhs), Value::String(rhs)) => Ok(Value::String(lhs + &rhs)),
-        (Value::String(lhs), rhs) => Ok(Value::String(lhs + &rhs.to_string())),
-        (lhs, Value::String(rhs)) => Ok(Value::String(lhs.to_string() + &rhs)),
-        (lhs, rhs) => match (numeric_value(&lhs), numeric_value(&rhs)) {
-            (Some(lhs), Some(rhs)) => {
-                let (lhs, rhs, kind) = promote_numeric_pair(lhs, rhs);
-                let value = match (lhs, rhs, kind) {
-                    (NumericValue::Int(lhs), NumericValue::Int(rhs), NumericKind::Int) => lhs
-                        .checked_add(rhs)
-                        .map(NumericValue::Int)
-                        .ok_or_else(|| Diagnostic::runtime(span, "integer overflow"))?,
-                    (NumericValue::Long(lhs), NumericValue::Long(rhs), NumericKind::Long) => lhs
-                        .checked_add(rhs)
-                        .map(NumericValue::Long)
-                        .ok_or_else(|| Diagnostic::runtime(span, "integer overflow"))?,
-                    (NumericValue::Float(lhs), NumericValue::Float(rhs), NumericKind::Float) => {
-                        NumericValue::Float(lhs + rhs)
-                    }
-                    (NumericValue::Double(lhs), NumericValue::Double(rhs), NumericKind::Double) => {
-                        NumericValue::Double(lhs + rhs)
-                    }
-                    _ => unreachable!("promoted numeric add should align kinds"),
-                };
-                Ok(value.into_value())
-            }
-            _ => Err(Diagnostic::runtime(
-                span,
-                "`+` expects numbers or strings of the same shape",
-            )),
-        },
-    }
-}
-
-fn eval_subtract(lhs: Value, rhs: Value, span: Span) -> Result<Value, Diagnostic> {
-    match (numeric_value(&lhs), numeric_value(&rhs)) {
-        (Some(lhs), Some(rhs)) => {
-            let (lhs, rhs, kind) = promote_numeric_pair(lhs, rhs);
-            let value = match (lhs, rhs, kind) {
-                (NumericValue::Int(lhs), NumericValue::Int(rhs), NumericKind::Int) => lhs
-                    .checked_sub(rhs)
-                    .map(NumericValue::Int)
-                    .ok_or_else(|| Diagnostic::runtime(span, "integer overflow"))?,
-                (NumericValue::Long(lhs), NumericValue::Long(rhs), NumericKind::Long) => lhs
-                    .checked_sub(rhs)
-                    .map(NumericValue::Long)
-                    .ok_or_else(|| Diagnostic::runtime(span, "integer overflow"))?,
-                (NumericValue::Float(lhs), NumericValue::Float(rhs), NumericKind::Float) => {
-                    NumericValue::Float(lhs - rhs)
-                }
-                (NumericValue::Double(lhs), NumericValue::Double(rhs), NumericKind::Double) => {
-                    NumericValue::Double(lhs - rhs)
-                }
-                _ => unreachable!("promoted numeric subtract should align kinds"),
-            };
-            Ok(value.into_value())
-        }
-        _ => Err(Diagnostic::runtime(span, "`-` expects numbers")),
-    }
-}
-
-fn eval_multiply(lhs: Value, rhs: Value, span: Span) -> Result<Value, Diagnostic> {
-    match (numeric_value(&lhs), numeric_value(&rhs)) {
-        (Some(lhs), Some(rhs)) => {
-            let (lhs, rhs, kind) = promote_numeric_pair(lhs, rhs);
-            let value = match (lhs, rhs, kind) {
-                (NumericValue::Int(lhs), NumericValue::Int(rhs), NumericKind::Int) => lhs
-                    .checked_mul(rhs)
-                    .map(NumericValue::Int)
-                    .ok_or_else(|| Diagnostic::runtime(span, "integer overflow"))?,
-                (NumericValue::Long(lhs), NumericValue::Long(rhs), NumericKind::Long) => lhs
-                    .checked_mul(rhs)
-                    .map(NumericValue::Long)
-                    .ok_or_else(|| Diagnostic::runtime(span, "integer overflow"))?,
-                (NumericValue::Float(lhs), NumericValue::Float(rhs), NumericKind::Float) => {
-                    NumericValue::Float(lhs * rhs)
-                }
-                (NumericValue::Double(lhs), NumericValue::Double(rhs), NumericKind::Double) => {
-                    NumericValue::Double(lhs * rhs)
-                }
-                _ => unreachable!("promoted numeric multiply should align kinds"),
-            };
-            Ok(value.into_value())
-        }
-        _ => Err(Diagnostic::runtime(span, "`*` expects numbers")),
-    }
-}
-
-fn eval_divide(lhs: Value, rhs: Value, span: Span) -> Result<Value, Diagnostic> {
-    match (numeric_value(&lhs), numeric_value(&rhs)) {
-        (Some(lhs), Some(rhs)) => {
-            let (lhs, rhs, kind) = promote_numeric_pair(lhs, rhs);
-            let value = match (lhs, rhs, kind) {
-                (_, NumericValue::Int(0), NumericKind::Int)
-                | (_, NumericValue::Long(0), NumericKind::Long)
-                | (_, NumericValue::Float(0.0), NumericKind::Float)
-                | (_, NumericValue::Double(0.0), NumericKind::Double) => {
-                    return Err(Diagnostic::runtime(span, "division by zero"));
-                }
-                (NumericValue::Int(lhs), NumericValue::Int(rhs), NumericKind::Int) => lhs
-                    .checked_div(rhs)
-                    .map(NumericValue::Int)
-                    .ok_or_else(|| Diagnostic::runtime(span, "integer overflow"))?,
-                (NumericValue::Long(lhs), NumericValue::Long(rhs), NumericKind::Long) => lhs
-                    .checked_div(rhs)
-                    .map(NumericValue::Long)
-                    .ok_or_else(|| Diagnostic::runtime(span, "integer overflow"))?,
-                (NumericValue::Float(lhs), NumericValue::Float(rhs), NumericKind::Float) => {
-                    NumericValue::Float(lhs / rhs)
-                }
-                (NumericValue::Double(lhs), NumericValue::Double(rhs), NumericKind::Double) => {
-                    NumericValue::Double(lhs / rhs)
-                }
-                _ => unreachable!("promoted numeric divide should align kinds"),
-            };
-            Ok(value.into_value())
-        }
-        _ => Err(Diagnostic::runtime(span, "`/` expects numbers")),
-    }
-}
-
-fn eval_comparison(op: BinaryOp, lhs: Value, rhs: Value, span: Span) -> Result<Value, Diagnostic> {
-    let (lhs, rhs) = match (numeric_value(&lhs), numeric_value(&rhs)) {
-        (Some(lhs), Some(rhs)) => {
-            let (lhs, rhs, _) = promote_numeric_pair(lhs, rhs);
-            (lhs.as_f64(), rhs.as_f64())
-        }
-        _ => return Err(Diagnostic::runtime(span, "comparison expects numbers")),
-    };
-    let result = match op {
-        BinaryOp::Less => lhs < rhs,
-        BinaryOp::LessEqual => lhs <= rhs,
-        BinaryOp::Greater => lhs > rhs,
-        BinaryOp::GreaterEqual => lhs >= rhs,
-        _ => unreachable!("comparison operator expected"),
-    };
-    Ok(Value::Bool(result))
-}
-
-fn eval_bitwise(op: BinaryOp, lhs: Value, rhs: Value, span: Span) -> Result<Value, Diagnostic> {
-    let (lhs, rhs) = match (lhs, rhs) {
-        (Value::Int(lhs), Value::Int(rhs)) => (lhs, rhs),
-        (Value::Long(lhs), Value::Long(rhs)) => {
-            let value = match op {
-                BinaryOp::BitAnd => lhs & rhs,
-                BinaryOp::BitOr => lhs | rhs,
-                BinaryOp::BitXor => lhs ^ rhs,
-                _ => unreachable!("bitwise operator expected"),
-            };
-            return Ok(Value::Long(value));
-        }
-        _ => {
-            return Err(Diagnostic::runtime(
-                span,
-                "bitwise operators expect integers",
-            ));
-        }
-    };
-    let value = match op {
-        BinaryOp::BitAnd => lhs & rhs,
-        BinaryOp::BitOr => lhs | rhs,
-        BinaryOp::BitXor => lhs ^ rhs,
-        _ => unreachable!("bitwise operator expected"),
-    };
-    Ok(Value::Int(value))
 }
 
 fn eval_call(
@@ -3468,192 +3006,6 @@ fn split_top_level(text: &str, separator: char) -> Vec<String> {
         parts.push(text[start..].trim().to_string());
     }
     parts.into_iter().filter(|part| !part.is_empty()).collect()
-}
-
-fn builtin_name(name: &str) -> Option<&'static str> {
-    match name {
-        "println" => Some("println"),
-        "printlnError" => Some("printlnError"),
-        "ToDo" => Some("ToDo"),
-        "assert" => Some("assert"),
-        "assertResult" => Some("assertResult"),
-        "thread" => Some("thread"),
-        "sleep" => Some("sleep"),
-        "stopwatch" => Some("stopwatch"),
-        "double" => Some("double"),
-        "int" => Some("int"),
-        "floor" => Some("floor"),
-        "ceil" => Some("ceil"),
-        "abs" => Some("abs"),
-        "sqrt" => Some("sqrt"),
-        "toString" => Some("toString"),
-        "substring" => Some("substring"),
-        "at" => Some("at"),
-        "matches" => Some("matches"),
-        "split" => Some("split"),
-        "join" => Some("join"),
-        "trim" => Some("trim"),
-        "trimLeft" => Some("trimLeft"),
-        "trimRight" => Some("trimRight"),
-        "replace" => Some("replace"),
-        "replaceAll" => Some("replaceAll"),
-        "toLowerCase" => Some("toLowerCase"),
-        "toUpperCase" => Some("toUpperCase"),
-        "startsWith" => Some("startsWith"),
-        "endsWith" => Some("endsWith"),
-        "contains" => Some("contains"),
-        "isEmptyString" => Some("isEmptyString"),
-        "indexOf" => Some("indexOf"),
-        "lastIndexOf" => Some("lastIndexOf"),
-        "length" => Some("length"),
-        "repeat" => Some("repeat"),
-        "reverse" => Some("reverse"),
-        "head" => Some("head"),
-        "tail" => Some("tail"),
-        "size" => Some("size"),
-        "isEmpty" => Some("isEmpty"),
-        "cons" => Some("cons"),
-        "map" => Some("map"),
-        "foldLeft" => Some("foldLeft"),
-        "FileInput#open" => Some("FileInput#open"),
-        "FileInput#readAll" => Some("FileInput#readAll"),
-        "FileInput#readLines" => Some("FileInput#readLines"),
-        "FileInput#all" => Some("FileInput#all"),
-        "FileInput#lines" => Some("FileInput#lines"),
-        "FileOutput#write" => Some("FileOutput#write"),
-        "FileOutput#append" => Some("FileOutput#append"),
-        "FileOutput#exists" => Some("FileOutput#exists"),
-        "FileOutput#delete" => Some("FileOutput#delete"),
-        "FileOutput#writeLines" => Some("FileOutput#writeLines"),
-        "Dir#current" => Some("Dir#current"),
-        "Dir#home" => Some("Dir#home"),
-        "Dir#temp" => Some("Dir#temp"),
-        "Dir#exists" => Some("Dir#exists"),
-        "Dir#mkdir" => Some("Dir#mkdir"),
-        "Dir#mkdirs" => Some("Dir#mkdirs"),
-        "Dir#isDirectory" => Some("Dir#isDirectory"),
-        "Dir#isFile" => Some("Dir#isFile"),
-        "Dir#list" => Some("Dir#list"),
-        "Dir#listFull" => Some("Dir#listFull"),
-        "Dir#delete" => Some("Dir#delete"),
-        "Dir#copy" => Some("Dir#copy"),
-        "Dir#move" => Some("Dir#move"),
-        "Map#containsKey" => Some("Map#containsKey"),
-        "Map#containsValue" => Some("Map#containsValue"),
-        "Map#get" => Some("Map#get"),
-        "Map#isEmpty" => Some("Map#isEmpty"),
-        "Map#size" => Some("Map#size"),
-        "Set#contains" => Some("Set#contains"),
-        "Set#isEmpty" => Some("Set#isEmpty"),
-        "Set#size" => Some("Set#size"),
-        _ => None,
-    }
-}
-
-fn builtin_arity(name: &str) -> Option<usize> {
-    match name {
-        "println" | "printlnError" | "assert" | "thread" | "sleep" | "stopwatch" | "double"
-        | "int" | "floor" | "ceil" | "abs" | "sqrt" | "toString" | "trim" | "trimLeft"
-        | "trimRight" | "toLowerCase" | "toUpperCase" | "isEmptyString" | "length" | "reverse"
-        | "head" | "tail" | "size" | "isEmpty" => Some(1),
-        "assertResult" | "at" | "matches" | "split" | "join" | "startsWith" | "endsWith"
-        | "contains" | "indexOf" | "lastIndexOf" | "repeat" | "cons" | "map" => Some(2),
-        "substring" | "replace" | "replaceAll" | "foldLeft" => Some(3),
-        "FileInput#readAll"
-        | "FileInput#readLines"
-        | "FileInput#all"
-        | "FileInput#lines"
-        | "FileOutput#exists"
-        | "FileOutput#delete"
-        | "Dir#exists"
-        | "Dir#mkdir"
-        | "Dir#mkdirs"
-        | "Dir#isDirectory"
-        | "Dir#isFile"
-        | "Dir#list"
-        | "Dir#listFull"
-        | "Dir#delete"
-        | "Map#isEmpty"
-        | "Map#size"
-        | "Set#isEmpty"
-        | "Set#size" => Some(1),
-        "FileInput#open"
-        | "FileOutput#write"
-        | "FileOutput#append"
-        | "FileOutput#writeLines"
-        | "Dir#copy"
-        | "Dir#move"
-        | "Map#containsKey"
-        | "Map#containsValue"
-        | "Map#get"
-        | "Set#contains" => Some(2),
-        "Dir#current" | "Dir#home" | "Dir#temp" | "ToDo" => Some(0),
-        _ => None,
-    }
-}
-
-fn value_method_builtin_name(value: &Value, field: &str) -> Option<&'static str> {
-    match value {
-        Value::Int(_)
-        | Value::Long(_)
-        | Value::Float(_)
-        | Value::Double(_)
-        | Value::Bool(_)
-        | Value::Unit
-        | Value::Record { .. } => match field {
-            "toString" => Some("toString"),
-            _ => None,
-        },
-        Value::String(_) => match field {
-            "toString" => Some("toString"),
-            "substring" => Some("substring"),
-            "at" => Some("at"),
-            "matches" => Some("matches"),
-            "split" => Some("split"),
-            "trim" => Some("trim"),
-            "trimLeft" => Some("trimLeft"),
-            "trimRight" => Some("trimRight"),
-            "replace" => Some("replace"),
-            "replaceAll" => Some("replaceAll"),
-            "toLowerCase" => Some("toLowerCase"),
-            "toUpperCase" => Some("toUpperCase"),
-            "startsWith" => Some("startsWith"),
-            "endsWith" => Some("endsWith"),
-            "contains" => Some("contains"),
-            "isEmpty" | "isEmptyString" => Some("isEmptyString"),
-            "indexOf" => Some("indexOf"),
-            "lastIndexOf" => Some("lastIndexOf"),
-            "length" => Some("length"),
-            "repeat" => Some("repeat"),
-            "reverse" => Some("reverse"),
-            _ => None,
-        },
-        Value::List(_) => match field {
-            "head" => Some("head"),
-            "tail" => Some("tail"),
-            "size" => Some("size"),
-            "isEmpty" => Some("isEmpty"),
-            "join" => Some("join"),
-            "map" => Some("map"),
-            "foldLeft" => Some("foldLeft"),
-            _ => None,
-        },
-        Value::Map(_) => match field {
-            "containsKey" => Some("Map#containsKey"),
-            "containsValue" => Some("Map#containsValue"),
-            "get" => Some("Map#get"),
-            "isEmpty" => Some("Map#isEmpty"),
-            "size" => Some("Map#size"),
-            _ => None,
-        },
-        Value::Set(_) => match field {
-            "contains" => Some("Set#contains"),
-            "isEmpty" => Some("Set#isEmpty"),
-            "size" => Some("Set#size"),
-            _ => None,
-        },
-        _ => None,
-    }
 }
 
 fn prefer_instance_dispatch(name: &str, argument_values: &[Value]) -> bool {
